requirements.txt
streamlit>=1.22.0
pandas>=1.5.0
openpyxl>=3.1.0
xlsxwriter>=3.0.0
openai>=1.3.0
anthropic>=0.8.0
python-dotenv>=1.0.0
requests>=2.28.0
streamlit-local-storage>=0.0.25